vscode app_py implementacao/requirements.txt
requests
python-nmap
bandit
dnspython
python-whois
